[refactor] validate 함수 utils 폴더 이동
